Moved file handling methods into API class
create storage file and read storage file methods moved into API. Added filelist that can be accessed by all API methods.

Devin McDermott <dc.mcdermott22@gmail.com>
