+++ src/main/java/org/jaropeners/App.java
@@ -7,11 +7,13 @@ import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
 
+import java.io.FileNotFoundException;
 import java.io.IOException;
 
 /**
  * JavaFX App
  */
+
 public class App extends Application {
 
     private static Scene scene;
@@ -35,8 +37,12 @@ public class App extends Application {
 
     }
 
-    public static void main(String[] args) {
+    public static void main(String[] args) throws FileNotFoundException {
+        // Prior to launching GUI
+        // Check for storage file and create it if doesn't exist
+        // If it does exist read the lines from the file and store them in the file List
+        API.createStorageFile();
+        API.readStorageFile();
         launch();
     }
-
 }
